backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.1
Pillow>=10.2.0
python-dotenv>=1.0.1
